backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
celery
redis
psycopg2-binary
python-dotenv
django-environ
gunicorn
uvicorn
structlog
python-json-logger
boto3
django-storages[boto3]
Pillow
daphne
channels
channels-redis
